A-Level Project for Computing
This is a GUI system designed to work as a point-of-sale tool for a
friend's boutique during A-Level. It was written in VB6. For simplicity,
please refer to the writing up documents for the GUI interface and the
actual functions of the system, as the actual system may require certain
installations to work.